+++ src/tr1/game/shell.c
@@ -200,18 +200,15 @@ void Shell_Main(void)
             gf_cmd.param);
 
         switch (gf_cmd.action) {
-        case GF_START_GAME: {
+        case GF_START_GAME:
+        case GF_SELECT_GAME: {
             const int32_t level_num = gf_cmd.param;
             const GF_LEVEL *const level = GF_GetLevel(GFLT_MAIN, level_num);
+            const GF_SEQUENCE_CONTEXT seq_ctx =
+                gf_cmd.action == GF_SELECT_GAME ? GFSC_SELECT : GFSC_NORMAL;
             if (level != nullptr) {
-                gf_cmd = GF_DoLevelSequence(level, GFSC_NORMAL);
+                gf_cmd = GF_DoLevelSequence(level, seq_ctx);
             }
-            break;
-        }
-
-        case GF_SELECT_GAME: {
-            const GF_LEVEL *const level = GF_GetLevel(GFLT_MAIN, gf_cmd.param);
-            gf_cmd = GF_DoLevelSequence(level, GFSC_SELECT);
             break;
         }