+++ docs/tr2/CHANGELOG.md
@@ -51,6 +51,7 @@
 - fixed software renderer skybox occlusion issues (#2343, regression from 0.7)
 - fixed gunflare from bandits in Tibetan levels spawning too far from their guns (#2365, regression from 0.8)
 - fixed guns sometimes appearing in Lara's hands when entering the fly cheat while undrawing weapons (#2376, regression from 0.3)
+- fixed the `/play` console command not resetting Lara's inventory (#2267, regression from 0.3)
 - improved rendering to achieve a slight performance boost in big rooms (#2325)
 - improved wireframe mode appearance around screen edges